moved homepage code into home.js
